+++ pricing_engine/api/v1/serializers/request/pricing_engine_request.py
@@ -22,7 +22,7 @@ class PricingEngineRequestSerializer(serializers.Serializer):
         write_only=True,
         allow_null=False,
     )
-    times_per_week = serializers.IntegerField(
+    times_per_week = serializers.DecimalField(
         required=False,
         write_only=True,
         allow_null=True,
